feat: Add setup task for Visual Studio Code
This commit adds a new setup task for Visual Studio Code in the `.vscode/tasks.json` file. The task runs the `setup-vscode.sh` script and has a Windows-specific command as well. Additionally, the commit includes the `setup-vscode.sh` script that installs a list of default extensions for Visual Studio Code, excluding one for Windows.

The added task automates the setup process, making it easier for users to get started with Visual Studio Code and ensures that the required extensions are installed.

Signed-off-by: Schubert Anselme <schubert@anselm.es>