chore: repoconfig
fix: Increase line length limit to 120

Increase the maximum line length to 120 characters in the .yamllint.yaml configuration file. This change is meant to enforce a consistent code style and improve code readability.

feat: Add vscode extensions config

Added a new extensions.json file to the .vscode directory. This file is used to configure the recommended and unwanted vscode extensions for the project. Currently, the file is empty, but it can be used to customize the vscode extensions based on project requirements.

refactor: Update setup-vscode.sh

In the setup-vscode.sh script, made modifications to the installation flow of vscode extensions. Now, the script installs all the default extensions defined in DEFAULT_EXTENSION_LIST, except for trunk.io, which is excluded for Windows systems. There is also an option to install additional recommended extensions by passing the "--recommendations" flag to the script. If there are no recommendations to install, the script displays a message and exits gracefully.

Signed-off-by: Schubert Anselme <schubert@anselm.es>