chore: repoconfig (#14)
fix: Increase line length limit to 120

Increase the maximum line length to 120 characters in the .yamllint.yaml
configuration file. This change is meant to enforce a consistent code
style and improve code readability.

feat: Add vscode extensions config

Added a new extensions.json file to the .vscode directory. This file is
used to configure the recommended and unwanted vscode extensions for the
project. Currently, the file is empty, but it can be used to customize
the vscode extensions based on project requirements.

refactor: Update setup-vscode.sh

In the setup-vscode.sh script, made modifications to the installation
flow of vscode extensions. Now, the script installs all the default
extensions defined in DEFAULT_EXTENSION_LIST, except for trunk.io, which
is excluded for Windows systems. There is also an option to install
additional recommended extensions by passing the "--recommendations"
flag to the script. If there are no recommendations to install, the
script displays a message and exits gracefully.

Signed-off-by: Schubert Anselme <schubert@anselm.es>

**Problem Statement**

Description of the issue

**Solution**

Description of changes that solved the issue

**Dependencies**

Other PRs that this PR depends on being merged at the same time

**Test Procedures**

How to test this PR

**Before**

How this looked/responded before the PR was created

**After**

How this looks/responds after the PR has been applied with a link to
test the PR

---

**Release notes**:

<!--  Steps to write your release note:
1. Use the release-note-* labels to set the release note state (if you
have access)
2. Enter your extended release note in the below block; leaving it blank
means using
the PR title as the release note. If no release note is required, just
write `NONE`.
-->

```release-note

```

Signed-off-by: Schubert Anselme <schubert@anselm.es>